[settings] Allow for background settings block update in the config UI
The config UI made the assumption that settings blocks were not updated in the
backgroud.  This update allows for settings addition and deletion while the user
is using the UI.  A message is displayed to let the user know updates took
place.  If the parrent settings block is removed, "../" will redirect to the
root settings block.

Signed-off-by: Marin Hannache <git@mareo.fr>